webPack

+++ src/controllers/videoController.js
@@ -1,10 +1,7 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 //Video.find({},(error,videos) => {});
-
-
-
-
 
 /*export const home = async(req,res) =>{
   try{
@@ -19,60 +16,83 @@ import Video from "../models/Video";
 */
 //callback
 
-export const home = async (req,res) =>{
-  const videos = await Video.find({}).sort({creatAt:"desc"});
+export const home = async (req, res) => {
+  const videos = await Video.find({}).sort({ creatAt: "desc" });
   console.log(videos);
-  return res.render("home",{pageTitle: "Home", videos});
-};//promise
+  return res.render("home", { pageTitle: "Home", videos });
+}; //promise
 
 export const watch = async (req, res) => {
-   const { id } = req.params;
-   const video = await Video.findById(id);
-   if (!video) {
+  const { id } = req.params;
+  const video = await Video.findById(id).populate("owner");
+  console.log("video", video);
+  if (!video) {
     return res.render("404", { pageTitle: "Video not found." });
   }
-   return res.render("watch", { pageTitle: video.title, video });
- };
-
-export const getEdit = async(req,res) =>{
-   const { id } = req.params;
-   const video = await Video.findById(id);
-   if (!video) {
-    return res.render("404", { pageTitle: "Video not found." });
-  }
-   return res.render("edit", { pageTitle: `Edit ${video.title}`,video});
+  return res.render("watch", { pageTitle: video.title, video });
 };
 
-export const postEdit = async(req,res) =>{
-  const {id} = req.params;
-  const {title, description, hashtags} = req.body;
-  const video = await Video.exists({_id: id});
+export const getEdit = async (req, res) => {
+  const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  return res.render("edit", { pageTitle: `Edit ${video.title}`, video });
+};
+
+export const postEdit = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+  const { id } = req.params;
+  const { title, description, hashtags } = req.body;
+  const video = await Video.exists({ _id: id });
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
   }
-  await Video.findByIdAndUpdate(id,{
-    title,description,hashtags : Video.formatHashtags(hashtags),
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  await Video.findByIdAndUpdate(id, {
+    title,
+    description,
+    hashtags: Video.formatHashtags(hashtags),
   });
 
   return res.redirect(`/videos/${id}`);
-
 };
 
-export const getUpload = (req,res) =>{
-  return res.render("upload",{pageTitle : "Upload Viedo"});
+export const getUpload = (req, res) => {
+  return res.render("upload", { pageTitle: "Upload Viedo" });
 };
-
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+  const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      fileUrl,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
+    console.log(error);
     return res.status(400).render("upload", {
       pageTitle: "Upload Video",
       errorMessage: error._message,
@@ -80,22 +100,31 @@ export const postUpload = async (req, res) => {
   }
 };
 
-
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
 
-export const search = async (req,res) => {
+export const search = async (req, res) => {
   const { keyword } = req.query;
   let videos = [];
-  if (keyword){
+  if (keyword) {
     videos = await Video.find({
       title: {
         $regex: new RegExp(`${keyword}$`, "i"),
       },
     });
   }
-  return res.render("search", {pageTitle:"Search", videos});
+  return res.render("search", { pageTitle: "Search", videos });
 };
